+++ app/src/main/AndroidManifest.xml
@@ -32,7 +32,7 @@
             android:exported="false" />
 
         <activity
-            android:name="com.seoulfitu.android.ui.facilityfilter.FacilityFilterActivity"
+            android:name="com.seoulfitu.android.ui.filter.facility.FacilityFilterActivity"
             android:exported="false" />
 
         <activity
