+++ app/src/main/java/com/seoulfitu/android/ui/facility/SportsFacilityBottomSheetFragment.kt
@@ -9,7 +9,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.activityViewModels
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.seoulfitu.android.databinding.FragmentSportsFacilityBottomSheetBinding
-import com.seoulfitu.android.ui.facilityfilter.FacilityFilterActivity
+import com.seoulfitu.android.ui.filter.facility.FacilityFilterActivity
 import com.seoulfitu.android.ui.uimodel.UiSelectedOptions
 import com.seoulfitu.android.util.getParcelableExtraCompat
 import dagger.hilt.android.AndroidEntryPoint
